requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.10
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0

# ClickHouse driver (FOUNDATION-0.3)
clickhouse-driver==0.2.9

# Qdrant vector database (FOUNDATION-0.4)
qdrant-client==1.15.1
